Move repeatability measurements into own file
