+++ python/lsst/validate/drp/calcnonsrd/__init__.py
@@ -1,4 +1,4 @@
 """Metric measurement classes *not* defined in LPM-17, the Science Requirements Document.
 """
 
-from .model_phot_rep import measureModelPhotRepeat # NOQA
+from .model_phot_rep import measure_model_phot_rep # NOQA
